:recycle: refactor : make client response wrapper package and rename response dtos
