+++ com.templlo.service.reservation/src/main/java/com/templlo/service/reservation/domain/reservation/client/UserClient.java
@@ -1,7 +1,7 @@
 package com.templlo.service.reservation.domain.reservation.client;
 
-import com.templlo.service.reservation.domain.reservation.client.model.response.GetUserByUserId;
-import com.templlo.service.reservation.domain.reservation.client.model.response.UserServiceWrapperRes;
+import com.templlo.service.reservation.domain.reservation.client.model.response.GetUserByUserIdRes;
+import com.templlo.service.reservation.domain.reservation.client.model.response.wrapper.UserServiceWrapperRes;
 import com.templlo.service.reservation.global.feign.AuthHeader;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,5 +13,5 @@ public interface UserClient {
 
     @AuthHeader
     @GetMapping("/api/users/{userId}")
-    UserServiceWrapperRes<GetUserByUserId> getUser(@PathVariable(name = "userId") String userId);
+    UserServiceWrapperRes<GetUserByUserIdRes> getUser(@PathVariable(name = "userId") String userId);
 }
