+++ com.templlo.service.reservation/src/main/java/com/templlo/service/reservation/domain/reservation/client/model/response/wrapper/UserServiceWrapperRes.java
@@ -1,4 +1,4 @@
-package com.templlo.service.reservation.domain.reservation.client.model.response;
+package com.templlo.service.reservation.domain.reservation.client.model.response.wrapper;
 
 public record UserServiceWrapperRes<T> (
         String status,
